refactor(coordinator/api): Phase 5.1 — split server.go into per-domain files
Pure-move split (deletion-only on the source; no body/comment/signature changed).
server.go 2043 -> 231 lines, keeping the Server struct, NewServer (init order
intact), and the api-key-cache consts.

New files: context_keys, server_setters, ratelimit_middleware, auth_middleware,
middleware (Handler/cors/recover/logging/statusWriter/newRequestID), routes
(with the //go:embed install.sh directive + installScript adjacency preserved),
runtime_manifest, binary_hash_policy, catalog_sync, observability, mdm_webhook.

Verified: gofmt clean, go build/vet, go test ./coordinator/api/... (throwaway
Postgres), and TestInstallScriptTemplating (install.sh embed) all green.
