+++ coordinator/api/mdm_webhook.go
@@ -1,0 +1,21 @@
+package api
+
+import (
+	"io"
+	"net/http"
+)
+
+// HandleMDMWebhook processes a MicroMDM webhook callback.
+// Mount this on the webhook URL configured in MicroMDM.
+func (s *Server) HandleMDMWebhook(w http.ResponseWriter, r *http.Request) {
+	body, err := io.ReadAll(r.Body)
+	if err != nil {
+		http.Error(w, "bad request", http.StatusBadRequest)
+		return
+	}
+	s.logger.Debug("mdm webhook received", "body_size", len(body), "body_preview", string(body[:min(len(body), 500)]))
+	if s.mdmClient != nil {
+		s.mdmClient.HandleWebhook(body)
+	}
+	w.WriteHeader(http.StatusOK)
+}
