+++ coordinator/api/observability.go
@@ -1,0 +1,140 @@
+package api
+
+import (
+	"context"
+	"net/http"
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+	"github.com/eigeninference/d-inference/coordinator/telemetry"
+)
+
+// emit is an internal convenience that funnels events through the emitter if
+// one has been wired up. No-op otherwise — telemetry must never affect control
+// flow.
+func (s *Server) emit(ctx context.Context, severity protocol.TelemetrySeverity, kind protocol.TelemetryKind, message string, fields map[string]any) {
+	if s.emitter == nil {
+		return
+	}
+	s.emitter.Emit(telemetry.Event{
+		Severity: severity,
+		Kind:     kind,
+		Message:  message,
+		Fields:   fields,
+	})
+}
+
+// emitRequest is like emit but preserves a request_id for correlation.
+func (s *Server) emitRequest(ctx context.Context, severity protocol.TelemetrySeverity, requestID, message string, fields map[string]any) {
+	if s.emitter == nil {
+		return
+	}
+	s.emitter.Emit(telemetry.Event{
+		Severity:  severity,
+		Kind:      protocol.KindInferenceError,
+		Message:   message,
+		Fields:    fields,
+		RequestID: requestID,
+	})
+}
+
+// ddIncr increments a DogStatsD counter. No-op if DD is not configured.
+func (s *Server) ddIncr(name string, tags []string) {
+	if s.dd != nil {
+		s.dd.Incr(name, tags)
+	}
+}
+
+// ddCount increments a DogStatsD counter by the given value. No-op if DD is not configured.
+func (s *Server) ddCount(name string, value int64, tags []string) {
+	if s.dd != nil {
+		s.dd.Count(name, value, tags)
+	}
+}
+
+// ddHistogram records a DogStatsD histogram value. No-op if DD is not configured.
+func (s *Server) ddHistogram(name string, value float64, tags []string) {
+	if s.dd != nil {
+		s.dd.Histogram(name, value, tags)
+	}
+}
+
+// ddGauge sets a DogStatsD gauge value. No-op if DD is not configured.
+func (s *Server) ddGauge(name string, value float64, tags []string) {
+	if s.dd != nil {
+		s.dd.Gauge(name, value, tags)
+	}
+}
+
+func (s *Server) emitPanic(ctx context.Context, message, stack string, fields map[string]any) {
+	if s.emitter == nil {
+		return
+	}
+	s.emitter.Emit(telemetry.Event{
+		Severity: protocol.SeverityFatal,
+		Kind:     protocol.KindPanic,
+		Message:  message,
+		Fields:   fields,
+		Stack:    stack,
+	})
+}
+
+// registerDefaultGauges wires live-computed gauges (fleet size, etc.) into
+// the metrics registry at construction time.
+func (s *Server) registerDefaultGauges() {
+	s.metrics.RegisterGauge("providers_online", func() float64 {
+		return float64(s.registry.ProviderCount())
+	})
+	s.metrics.RegisterGauge("min_provider_version_set", func() float64 {
+		if s.minProviderVersion != "" {
+			return 1
+		}
+		return 0
+	})
+}
+
+// StartDDGaugeLoop periodically pushes gauge values to DogStatsD. Gauges
+// are point-in-time values and must be pushed regularly (not on-demand like
+// counters). Call as a goroutine; stops when ctx is cancelled.
+func (s *Server) StartDDGaugeLoop(ctx context.Context) {
+	if s.dd == nil {
+		return
+	}
+	ticker := time.NewTicker(15 * time.Second)
+	defer ticker.Stop()
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case <-ticker.C:
+			s.ddGauge("providers.online", float64(s.registry.OnlineCount()), nil)
+			for model, count := range s.registry.ModelProviderSnapshot() {
+				s.ddGauge("providers.per_model", float64(count), []string{"model:" + model})
+			}
+			for ver, count := range s.registry.ProviderCountByVersion() {
+				s.ddGauge("providers.per_version", float64(count), []string{"version:" + ver})
+			}
+			if s.minProviderVersion != "" {
+				s.ddGauge("coordinator.min_provider_version_set", 1, []string{"min_version:" + s.minProviderVersion})
+			}
+			if q := s.registry.Queue(); q != nil {
+				s.ddGauge("request_queue.depth", float64(q.TotalSize()), nil)
+			}
+		}
+	}
+}
+
+// handleAdminMetrics returns the metrics snapshot in JSON or Prometheus text.
+func (s *Server) handleAdminMetrics(w http.ResponseWriter, r *http.Request) {
+	if !s.isAdminAuthorized(w, r) {
+		return
+	}
+	snap := s.metrics.Snapshot()
+	if r.URL.Query().Get("format") == "prom" {
+		w.Header().Set("Content-Type", "text/plain; version=0.0.4")
+		w.WriteHeader(http.StatusOK)
+		_, _ = w.Write([]byte(snap.RenderProm()))
+		return
+	}
+	writeJSON(w, http.StatusOK, snap)
+}
