+++ coordinator/api/routes.go
@@ -1,0 +1,251 @@
+package api
+
+import (
+	_ "embed"
+	"fmt"
+	"io"
+	"net/http"
+	"strings"
+)
+
+//go:embed install.sh
+var installScript []byte
+
+// installScriptPlaceholder is substituted with the coordinator's public URL at
+// serve time. Keep in sync with coordinator/internal/api/install.sh.
+//
+// The legacy install.sh also substituted __DARKBLOOM_R2_CDN_URL__ and
+// __DARKBLOOM_R2_SITE_PACKAGES_CDN_URL__ for the Python runtime download.
+// Post-Swift-cutover (v0.5.0+) install.sh no longer touches R2 directly --
+// model downloads run inside `darkbloom models download` against the public
+// R2 CDN -- so only the coordinator URL needs serve-time templating.
+const installScriptPlaceholder = "__DARKBLOOM_COORD_URL__"
+
+// resolveBaseURL returns the configured baseURL, or derives one from the
+// request's Host header when baseURL is unset. TLS-terminating proxies pass
+// through the original scheme via X-Forwarded-Proto; default to https.
+func (s *Server) resolveBaseURL(r *http.Request) string {
+	if s.baseURL != "" {
+		return s.baseURL
+	}
+	scheme := "https"
+	if proto := r.Header.Get("X-Forwarded-Proto"); proto != "" {
+		scheme = proto
+	} else if r.TLS == nil {
+		scheme = "http"
+	}
+	return scheme + "://" + r.Host
+}
+
+// routes mounts all HTTP and WebSocket handlers.
+func (s *Server) routes() {
+	// Install script — served from embedded binary with coordinator URL +
+	// R2 CDN URLs substituted per environment.
+	s.mux.HandleFunc("GET /install.sh", func(w http.ResponseWriter, r *http.Request) {
+		rendered := strings.ReplaceAll(string(installScript), installScriptPlaceholder, s.resolveBaseURL(r))
+		w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+		w.Header().Set("Cache-Control", "no-cache")
+		io.WriteString(w, rendered)
+	})
+
+	// Health check — no auth required.
+	s.mux.HandleFunc("GET /health", s.handleHealth)
+
+	// Provider WebSocket — no API key auth (providers authenticate differently).
+	s.mux.HandleFunc("GET /ws/provider", s.handleProviderWS)
+
+	// Key management — requires interactive Privy session (API keys rejected
+	// to prevent self-replication from a leaked key).
+	s.mux.HandleFunc("POST /v1/auth/keys", s.requirePrivyAuth(s.rateLimitFinancial(s.handleCreateKey)))
+	s.mux.HandleFunc("DELETE /v1/auth/keys", s.requirePrivyAuth(s.handleRevokeKey))
+
+	// Multi-key management (OpenRouter-shaped CRUD). One account may own many
+	// named, individually-limited keys. Management requires an interactive
+	// Privy session so a leaked inference key can't enumerate or mint keys.
+	s.mux.HandleFunc("GET /v1/keys", s.requirePrivyAuth(s.handleListAPIKeys))
+	s.mux.HandleFunc("POST /v1/keys", s.requirePrivyAuth(s.rateLimitFinancial(s.handleCreateAPIKey)))
+	s.mux.HandleFunc("GET /v1/keys/{id}", s.requirePrivyAuth(s.handleGetAPIKey))
+	s.mux.HandleFunc("PATCH /v1/keys/{id}", s.requirePrivyAuth(s.rateLimitFinancial(s.handleUpdateAPIKey)))
+	s.mux.HandleFunc("DELETE /v1/keys/{id}", s.requirePrivyAuth(s.rateLimitFinancial(s.handleDeleteAPIKey)))
+	s.mux.HandleFunc("POST /v1/keys/{id}/rotate", s.requirePrivyAuth(s.rateLimitFinancial(s.handleRotateAPIKey)))
+	// Metadata for the calling key (OpenRouter parity) — API key auth.
+	s.mux.HandleFunc("GET /v1/key", s.requireAuth(s.handleGetCallingKey))
+
+	// Consumer endpoints — API key auth required + per-account rate limit.
+	// Inference endpoints are wrapped in sealedTransport so senders can opt into
+	// sender→coordinator encryption by setting Content-Type:
+	// application/eigeninference-sealed+json (see sender_encryption.go).
+	// rateLimitConsumer is chained inside requireAuth so the accountID is in
+	// context. Read-only endpoints (GET /v1/models) skip rate limiting since
+	// they're cheap and clients poll them.
+	s.mux.HandleFunc("POST /v1/chat/completions", s.requireAuth(s.rateLimitConsumer(s.sealedTransport(s.handleChatCompletions))))
+	s.mux.HandleFunc("POST /v1/responses", s.requireAuth(s.rateLimitConsumer(s.sealedTransport(s.handleChatCompletions)))) // Responses API — same handler, auto-detects input vs messages
+	s.mux.HandleFunc("POST /v1/completions", s.requireAuth(s.rateLimitConsumer(s.sealedTransport(s.handleCompletions))))
+	s.mux.HandleFunc("POST /v1/messages", s.requireAuth(s.rateLimitConsumer(s.sealedTransport(s.handleAnthropicMessages))))
+	s.mux.HandleFunc("GET /v1/models", s.requireAuth(s.handleListModels))
+	// Dedicated OpenRouter provider feed — pure OpenRouter schema, no Darkbloom metadata.
+	s.mux.HandleFunc("GET /v1/models/openrouter", s.requireAuth(s.handleListModelsOpenRouter))
+
+	// Sender encryption — public key publication for sender→coordinator E2E.
+	// Optional: senders may use this to encrypt request bodies; plaintext path
+	// continues to work unchanged when this header isn't set.
+	s.mux.HandleFunc("GET /v1/encryption-key", s.handleEncryptionKey)
+
+	// MDM webhook — MicroMDM sends command responses here.
+	s.mux.HandleFunc("POST /v1/mdm/webhook", s.HandleMDMWebhook)
+
+	// Payment endpoints — API key auth required.
+	s.mux.HandleFunc("GET /v1/payments/balance", s.requireAuth(s.handleBalance))
+	s.mux.HandleFunc("GET /v1/payments/usage", s.requireAuth(s.handleUsage))
+
+	// Provider earnings — no API key auth (providers identify by provider address).
+	s.mux.HandleFunc("GET /v1/provider/earnings", s.handleProviderEarnings)
+
+	// Per-node provider earnings — public by provider_key, or auth'd by account.
+	s.mux.HandleFunc("GET /v1/provider/node-earnings", s.handleNodeEarnings)
+	s.mux.HandleFunc("GET /v1/provider/account-earnings", s.requireAuth(s.handleAccountEarnings))
+
+	// Account-scoped provider dashboard.
+	s.mux.HandleFunc("GET /v1/me/providers", s.requirePrivyAuth(s.handleMyProviders))
+	s.mux.HandleFunc("GET /v1/me/summary", s.requirePrivyAuth(s.handleMySummary))
+
+	// ACME enrollment — generates per-device .mobileconfig for device-attest-01.
+	// No auth needed — security comes from Apple's attestation during ACME challenge.
+	s.mux.HandleFunc("POST /v1/enroll", s.handleEnroll)
+
+	// Attestation verification — public, no auth needed.
+	// Users can independently verify Apple's MDA certificate chain.
+	s.mux.HandleFunc("GET /v1/providers/attestation", s.handleProviderAttestation)
+
+	// Capacity snapshot — no auth needed. Upstream routers poll this.
+	s.mux.HandleFunc("GET /v1/models/capacity", s.handleModelsCapacity)
+
+	// Platform stats — no auth needed. Frontend dashboard uses this.
+	s.mux.HandleFunc("GET /v1/stats", s.handleStats)
+
+	// Public leaderboard + network totals — no auth, pseudonymized,
+	// 5-min/1-min cache.
+	s.mux.HandleFunc("GET /v1/leaderboard", s.handleLeaderboard)
+	s.mux.HandleFunc("GET /v1/network/totals", s.handleNetworkTotals)
+
+	// Provider version check — no auth needed. Providers call this to check for updates.
+	s.mux.HandleFunc("GET /api/version", s.handleVersion)
+
+	// Releases — versioned provider binary distribution.
+	s.mux.HandleFunc("POST /v1/releases", s.handleRegisterRelease)     // scoped release key (GitHub Action)
+	s.mux.HandleFunc("GET /v1/releases/latest", s.handleLatestRelease) // public (install.sh)
+
+	// Device authorization flow — providers link to user accounts.
+	s.mux.HandleFunc("POST /v1/device/code", s.handleDeviceCode)   // no auth — provider not yet authenticated
+	s.mux.HandleFunc("POST /v1/device/token", s.handleDeviceToken) // no auth — polls with device_code secret
+	// Device approve issues a long-lived provider→account linking token —
+	// same risk class as /v1/auth/keys, so financial-tier limit applies.
+	// Uses requirePrivyAuth to reject API keys (interactive session only).
+	s.mux.HandleFunc("POST /v1/device/approve", s.requirePrivyAuth(s.rateLimitFinancial(s.handleDeviceApprove)))
+
+	// --- Billing endpoints (Stripe payments + referrals) ---
+
+	// Stripe — financial limiter on session creation (creates a checkout
+	// intent, hits external API). Read-only status endpoint not throttled.
+	s.mux.HandleFunc("POST /v1/billing/stripe/create-session", s.requireAuth(s.rateLimitFinancial(s.handleStripeCreateSession)))
+	s.mux.HandleFunc("POST /v1/billing/stripe/webhook", s.handleStripeWebhook) // no auth — Stripe signs it
+	s.mux.HandleFunc("GET /v1/billing/stripe/session", s.requireAuth(s.handleStripeSessionStatus))
+
+	// Wallet balance
+	s.mux.HandleFunc("GET /v1/billing/wallet/balance", s.requireAuth(s.handleWalletBalance))
+
+	// Stripe Payouts (Connect Express) — bank/card withdrawals.
+	s.mux.HandleFunc("POST /v1/billing/stripe/onboard", s.requireAuth(s.handleStripeOnboard))
+	s.mux.HandleFunc("GET /v1/billing/stripe/status", s.requireAuth(s.handleStripeStatus))
+	s.mux.HandleFunc("POST /v1/billing/withdraw/stripe", s.requireAuth(s.handleStripeWithdraw))
+	s.mux.HandleFunc("GET /v1/billing/stripe/withdrawals", s.requireAuth(s.handleStripeWithdrawals))
+	s.mux.HandleFunc("POST /v1/billing/stripe/connect/webhook", s.handleStripeConnectWebhook) // no auth — Stripe signs it
+
+	// Pricing — GET is public, PUT/DELETE require auth
+	s.mux.HandleFunc("GET /v1/pricing", s.handleGetPricing)                        // public
+	s.mux.HandleFunc("PUT /v1/pricing", s.requireAuth(s.handleSetPricing))         // provider sets own prices
+	s.mux.HandleFunc("DELETE /v1/pricing", s.requireAuth(s.handleDeletePricing))   // revert to default
+	s.mux.HandleFunc("PUT /v1/admin/pricing", s.requireAuth(s.handleAdminPricing)) // platform sets defaults
+
+	// Admin account management (service-role + per-account platform fee)
+	s.mux.HandleFunc("PUT /v1/admin/users/role", s.requireAuth(s.handleAdminSetUserRole))
+	s.mux.HandleFunc("PUT /v1/admin/users/platform-fee", s.requireAuth(s.handleAdminSetUserPlatformFee))
+
+	// Admin model registry (manifest-backed). The legacy supported_models CRUD
+	// (bare GET/POST/DELETE /v1/admin/models) was removed; the model_registry is
+	// the single source of truth. Use register + the per-model action endpoints.
+	s.mux.HandleFunc("POST /v1/admin/models/register", s.handleRegisterModel)
+	s.mux.HandleFunc("POST /v1/admin/models/", s.handleAdminModelRegistryAction)
+	s.mux.HandleFunc("GET /v1/admin/releases", s.handleAdminListReleases)     // admin key or Privy admin
+	s.mux.HandleFunc("DELETE /v1/admin/releases", s.handleAdminDeleteRelease) // admin key or Privy admin
+
+	// Admin CLI auth — Privy email OTP for getting admin tokens without a browser.
+	s.mux.HandleFunc("POST /v1/admin/auth/init", s.handleAdminAuthInit)     // no auth (sends OTP)
+	s.mux.HandleFunc("POST /v1/admin/auth/verify", s.handleAdminAuthVerify) // no auth (returns token)
+
+	// Public model catalog — providers and install script fetch this
+	s.mux.HandleFunc("GET /v1/models/catalog", s.handleModelCatalog)
+	s.mux.HandleFunc("GET /v1/models/catalog/manifest/", s.handleModelCatalogManifest)
+	s.mux.HandleFunc("GET /v1/models/catalog/", s.handleModelCatalogItem)
+
+	// Runtime manifest — providers and users can inspect accepted runtime hashes.
+	s.mux.HandleFunc("GET /v1/runtime/manifest", s.handleRuntimeManifest)
+
+	// Payment methods info
+	s.mux.HandleFunc("GET /v1/billing/methods", s.handleBillingMethods) // no auth needed
+
+	// Referral system — register/apply mutate referral graph (financial
+	// limiter); stats/info are read-only.
+	s.mux.HandleFunc("POST /v1/referral/register", s.requireAuth(s.rateLimitFinancial(s.handleReferralRegister)))
+	s.mux.HandleFunc("POST /v1/referral/apply", s.requireAuth(s.rateLimitFinancial(s.handleReferralApply)))
+	s.mux.HandleFunc("GET /v1/referral/stats", s.requireAuth(s.handleReferralStats))
+	s.mux.HandleFunc("GET /v1/referral/info", s.requireAuth(s.handleReferralInfo))
+
+	// Invite codes (admin)
+	// Invite code creation accepts amount_usd and produces a credit-bearing
+	// code; redemption is already financial-tier so the issuance side must
+	// match (otherwise an admin-key holder could spam codes anyway, but
+	// keeping symmetry).
+	s.mux.HandleFunc("POST /v1/admin/invite-codes", s.requireAuth(s.rateLimitFinancial(s.handleAdminCreateInviteCode)))
+	s.mux.HandleFunc("GET /v1/admin/invite-codes", s.requireAuth(s.handleAdminListInviteCodes))
+	s.mux.HandleFunc("DELETE /v1/admin/invite-codes", s.requireAuth(s.handleAdminDeactivateInviteCode))
+
+	// Invite code redemption (user) — credits the redeemer's balance, so
+	// it's a financial-tier endpoint.
+	s.mux.HandleFunc("POST /v1/invite/redeem", s.requireAuth(s.rateLimitFinancial(s.handleRedeemInviteCode)))
+
+	// Admin credit & reward
+	s.mux.HandleFunc("POST /v1/admin/credit", s.requireAuth(s.handleAdminCredit))
+	s.mux.HandleFunc("POST /v1/admin/reward", s.requireAuth(s.handleAdminReward))
+
+	// Telemetry ingestion — authentication is resolved inside the handler
+	// because providers, consumers, and anonymous clients all hit this path.
+	// Events are forwarded to Datadog; admin read/summary endpoints have been
+	// removed (use Datadog Log Explorer).
+	s.mux.HandleFunc("POST /v1/telemetry/events", s.handleTelemetryIngest)
+
+	// Provider log reports
+	s.mux.HandleFunc("POST /v1/provider/log-report", s.requireAuth(s.handleUploadLogReport))
+	s.mux.HandleFunc("GET /v1/admin/log-reports", s.requireAuth(s.handleListLogReports))
+	s.mux.HandleFunc("GET /v1/admin/log-reports/{id}", s.requireAuth(s.handleGetLogReport))
+
+	// Metrics snapshot (admin only)
+	s.mux.HandleFunc("GET /v1/admin/metrics", s.handleAdminMetrics)
+
+	// Catch-all for unimplemented OpenAI-compatible endpoints.
+	// Registered last (old-style pattern) so explicit method+path routes
+	// take precedence. Any /v1/* path not handled above gets a structured
+	// JSON error instead of the mux default text/plain 404.
+	s.mux.HandleFunc("/v1/", s.handleUnimplementedEndpoint)
+}
+
+// handleUnimplementedEndpoint returns a structured JSON error for any /v1/*
+// path not registered as an explicit route. This prevents OpenAI SDK clients
+// from crashing on raw text/plain 404s when hitting unimplemented endpoints
+// like /v1/embeddings or /v1/moderations.
+func (s *Server) handleUnimplementedEndpoint(w http.ResponseWriter, r *http.Request) {
+	writeJSON(w, http.StatusNotFound, errorResponse(
+		"invalid_request_error",
+		fmt.Sprintf("endpoint %s %s is not implemented", r.Method, r.URL.Path),
+	))
+}
