+++ coordinator/api/server_setters.go
@@ -1,0 +1,168 @@
+package api
+
+import (
+	"crypto/x509"
+	"strings"
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/auth"
+	"github.com/eigeninference/d-inference/coordinator/billing"
+	"github.com/eigeninference/d-inference/coordinator/datadog"
+	"github.com/eigeninference/d-inference/coordinator/internal/e2e"
+	"github.com/eigeninference/d-inference/coordinator/mdm"
+	"github.com/eigeninference/d-inference/coordinator/ratelimit"
+	"github.com/eigeninference/d-inference/coordinator/telemetry"
+)
+
+// SetRateLimiter configures the per-account rate limiter applied to
+// consumer inference endpoints. Pass nil to disable.
+func (s *Server) SetRateLimiter(rl *ratelimit.Limiter) {
+	s.rateLimiter = rl
+}
+
+// SetFinancialRateLimiter configures a stricter per-account limiter for
+// balance-mutating endpoints. Pass nil to disable.
+func (s *Server) SetFinancialRateLimiter(rl *ratelimit.Limiter) {
+	s.financialRateLimiter = rl
+}
+
+// SetServiceRateLimiter configures the elevated limiter used for service-role
+// accounts (e.g. OpenRouter). Pass nil to let service accounts bypass limits.
+func (s *Server) SetServiceRateLimiter(rl *ratelimit.Limiter) {
+	s.serviceRateLimiter = rl
+}
+
+// SetTokenLimiters configures the per-account input/output token-per-minute
+// limiters for the consumer and service tiers. Pass nil for a tier to disable
+// token limiting for it.
+func (s *Server) SetTokenLimiters(consumer, service *ratelimit.TokenLimiter) {
+	s.consumerTokenLimiter = consumer
+	s.serviceTokenLimiter = service
+}
+
+// SetKeyLimiters configures the per-key (variable-rate) RPM and ITPM/OTPM
+// limiters used for per-key overrides. Pass nil to disable per-key limiting.
+func (s *Server) SetKeyLimiters(rpm *ratelimit.Limiter, tokens *ratelimit.KeyTokenLimiter) {
+	s.keyRPMLimiter = rpm
+	s.keyTokenLimiter = tokens
+}
+
+// SetAdminKey configures the admin API key for admin-only endpoints.
+func (s *Server) SetAdminKey(key string) {
+	s.adminKey = key
+}
+
+// SetMinProviderVersion sets the minimum provider version for routing.
+func (s *Server) SetMinProviderVersion(v string) {
+	s.minProviderVersion = strings.TrimSpace(v)
+}
+
+// SetBaseURL sets the coordinator's public URL (used to template install.sh).
+// Pass the canonical origin with no trailing slash, e.g. "https://api.darkbloom.dev".
+// If unset, the install.sh handler derives a URL from the request's Host header.
+func (s *Server) SetBaseURL(url string) {
+	s.baseURL = strings.TrimRight(url, "/")
+}
+
+// SetR2CDNURL sets the public R2 bucket URL that install.sh substitutes as
+// the model/template/release download origin. If unset, install.sh keeps the
+// placeholder — providers will fail to pull artifacts, making the misconfig
+// loud instead of silent.
+func (s *Server) SetR2CDNURL(url string) {
+	s.r2CDNURL = strings.TrimRight(url, "/")
+}
+
+// SetEmitter wires the coordinator-side telemetry emitter. Call once at boot.
+func (s *Server) SetEmitter(e *telemetry.Emitter) {
+	s.emitter = e
+}
+
+// SetDatadog wires the Datadog client for DogStatsD metrics and Logs API forwarding.
+func (s *Server) SetDatadog(dd *datadog.Client) {
+	s.dd = dd
+}
+
+// Datadog returns the Datadog client (or nil). Exposed so main.go and the
+// telemetry emitter can share the same client.
+func (s *Server) Datadog() *datadog.Client {
+	return s.dd
+}
+
+// Metrics returns the in-process metrics registry so cmd/coordinator can
+// expose it to the telemetry emitter and other integrations.
+func (s *Server) Metrics() *Metrics {
+	return s.metrics
+}
+
+// SetStepCACerts configures the step-ca CA certificates for ACME client cert verification.
+func (s *Server) SetStepCACerts(root, intermediate *x509.Certificate) {
+	s.stepCARootCert = root
+	s.stepCAIntermediateCert = intermediate
+}
+
+// SetBilling configures the billing service for multi-chain payments and referrals.
+func (s *Server) SetBilling(svc *billing.Service) {
+	s.billing = svc
+}
+
+func (s *Server) Billing() *billing.Service {
+	return s.billing
+}
+
+func (s *Server) SetChallengeInterval(d time.Duration) {
+	s.challengeInterval = d
+}
+
+func (s *Server) SetSkipChallenge(skip bool) {
+	s.skipChallenge = skip
+}
+
+// SetPrivyAuth configures Privy JWT authentication for consumer endpoints.
+func (s *Server) SetPrivyAuth(pa *auth.PrivyAuth) {
+	s.privyAuth = pa
+}
+
+// SetAdminEmails configures which Privy accounts have admin access.
+func (s *Server) SetAdminEmails(emails []string) {
+	s.adminEmails = make(map[string]bool, len(emails))
+	for _, e := range emails {
+		s.adminEmails[strings.ToLower(strings.TrimSpace(e))] = true
+	}
+}
+
+// SetMDMClient configures the MicroMDM client for provider verification.
+// When set, providers are verified against MDM on registration.
+func (s *Server) SetMDMClient(client *mdm.Client) {
+	s.mdmClient = client
+}
+
+// SetConsoleURL sets the frontend URL for device auth verification links.
+func (s *Server) SetConsoleURL(url string) {
+	s.consoleURL = url
+}
+
+// SetCORSOrigin configures the allowed CORS origin.
+func (s *Server) SetCORSOrigin(origin string) {
+	s.corsOrigin = origin
+}
+
+// SetReleaseKey configures the scoped release key for GitHub Actions.
+func (s *Server) SetReleaseKey(key string) {
+	s.releaseKey = key
+}
+
+// SetCoordinatorKey installs the X25519 keypair the coordinator publishes
+// for sender-to-coordinator request encryption. Pass nil to disable.
+func (s *Server) SetCoordinatorKey(k *e2e.CoordinatorKey) {
+	s.coordinatorKey = k
+}
+
+// CoordinatorKey returns the configured coordinator encryption key (or nil).
+// Exposed for tests; production code should not need this.
+func (s *Server) CoordinatorKey() *e2e.CoordinatorKey {
+	return s.coordinatorKey
+}
+
+func (s *Server) SetRuntimeManifest(m *RuntimeManifest) {
+	s.knownRuntimeManifest = m
+}
